found the source of the issue, putting a panic in there while I fix it

+++ kernel/src/include/arch/arch_smp.h
@@ -5,6 +5,6 @@
 #ifndef KERNEL_ARCH_SMP_H
 #define KERNEL_ARCH_SMP_H
 void arch_smp_query();
-extern uint64 bootstrap_lapic_id;
+extern uint32 bootstrap_lapic_id;
 extern uint64 cpu_count;
 #endif //KERNEL_ARCH_SMP_H
